[dewbee-004] develop v1.3.3 1. `사이트 성능` 최적화  - 이미지 재조정 2. `패치내역` 조회방식 변경  - 서버측에서 조회처리
