+++ app/components/main/intro/Descriptions.vue
@@ -81,6 +81,8 @@ const { windowSize } = storeToRefs(useWindowStore())
         :src="url(true, '/assets/dewbee-logo.webp')"
         class="rounded-full ring-4 ring-amber-500"
         sizes="160px xs:300px"
+        :width="windowSize > 340 ? 300 : 160"
+        :height="windowSize > 340 ? 300 : 160"
         loading="lazy"
         placeholder
         format="webp"
